feat: filters, export, pagination, visuals
