+++ libs/oeth/history/src/components/woeth/HistoryCard/HistoryTable.tsx
@@ -1,93 +1,92 @@
-import { useMemo } from 'react';
-
 import {
   Box,
-  Paper,
+  Link,
   Stack,
   Table,
   TableBody,
   TableCell,
-  TableContainer,
   TableHead,
   TableRow,
   Typography,
 } from '@mui/material';
-import { ExternalLink, TransactionIcon } from '@origin/shared/components';
+import { TablePagination, TransactionIcon } from '@origin/shared/components';
 import { getTokenByAddress } from '@origin/shared/contracts';
+import { FaArrowUpRightRegular } from '@origin/shared/icons';
 import { getChain } from '@origin/shared/providers';
 import { formatAmount, txLink } from '@origin/shared/utils';
-import { indexBy } from 'ramda';
+import {
+  getCoreRowModel,
+  getPaginationRowModel,
+  useReactTable,
+} from '@tanstack/react-table';
 import { useIntl } from 'react-intl';
-import { useAccount, useConfig } from 'wagmi';
+import { useConfig } from 'wagmi';
 
 import type { HistoryType } from '@origin/oeth/shared';
 
-import type { BalancesQuery, TransfersQuery } from '../../../queries.generated';
-
 export type HistoryTableProps = {
-  transfers: TransfersQuery['erc20Transfers'];
-  balances: BalancesQuery['erc20Balances'];
+  rows: {
+    id: string;
+    chainId: number;
+    blockNumber: number;
+    timestamp: string;
+    address: string;
+    txHash: string;
+    type: HistoryType;
+    change: string;
+    balance: string;
+  }[];
 };
 
-export function HistoryTable({ transfers, balances }: HistoryTableProps) {
+export function HistoryTable({ rows }: HistoryTableProps) {
   const config = useConfig();
   const intl = useIntl();
-  const { address: userAddress } = useAccount();
-  const rows = useMemo(() => {
-    if (!userAddress) return [];
-    const balanceMap = indexBy(
-      (b) =>
-        `${b.chainId}-${b.blockNumber}-${b.address}-${b.account.toLowerCase()}`,
-      balances,
-    );
-    return transfers.map((t) => {
-      return {
-        id: t.id,
-        chainId: t.chainId,
-        blockNumber: t.blockNumber,
-        timestamp: t.timestamp,
-        address: t.address,
-        txHash: t.txHash,
-        type: (t.from.toLowerCase() === userAddress.toLowerCase()
-          ? 'Sent'
-          : 'Received') as HistoryType,
-        change: t.value,
-        balance:
-          balanceMap[
-            `${t.chainId}-${t.blockNumber}-${t.address}-${userAddress.toLowerCase()}`
-          ]?.balance,
-      };
-    });
-  }, [userAddress, transfers, balances]);
+  const table = useReactTable({
+    columns: [{ id: 'type' }, { id: 'change' }, { id: 'balance' }],
+    data: rows,
+    initialState: {
+      pagination: {
+        pageIndex: 0,
+        pageSize: 10,
+      },
+    },
+    getCoreRowModel: getCoreRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    autoResetPageIndex: true,
+  });
+  const displayRows = table.getRowModel().rows;
+
+  const cellSx = { paddingX: { xs: 1, md: 3 } };
 
   return (
-    <TableContainer component={Paper}>
+    <Stack>
       <Table>
         <TableHead>
           <TableRow>
-            <TableCell>Type</TableCell>
-            <TableCell align="right">Change</TableCell>
-            <TableCell align="right">Balance</TableCell>
-            <TableCell width={1} sx={{ paddingLeft: 0 }} />
+            <TableCell sx={{ py: 3, width: '50%', ...cellSx }}>Type</TableCell>
+            <TableCell align="right" sx={cellSx}>
+              Change
+            </TableCell>
+            <TableCell align="right" sx={cellSx}>
+              Balance
+            </TableCell>
+            <TableCell />
           </TableRow>
         </TableHead>
         <TableBody>
-          {rows.map((row) => {
+          {displayRows.map(({ original: row }) => {
             const token = getTokenByAddress(row.address, row.chainId);
             if (!token) return null;
             return (
-              <TableRow
-                key={row.id}
-                sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
-              >
-                <TableCell component="th" scope="row">
+              <TableRow key={row.id}>
+                <TableCell component="th" scope="row" sx={cellSx}>
                   <Stack direction={'row'} alignItems={'center'} gap={2}>
                     <TransactionIcon
                       type={row.type}
                       token={token}
                       sx={{
-                        width: 32,
-                        height: 32,
+                        width: { xs: 24, sm: 32 },
+                        height: { xs: 24, sm: 32 },
                       }}
                     />
                     <Box>
@@ -98,20 +97,30 @@ export function HistoryTable({ transfers, balances }: HistoryTableProps) {
                     </Box>
                   </Stack>
                 </TableCell>
-                <TableCell align="right">{formatAmount(row.change)}</TableCell>
-                <TableCell align="right">
+                <TableCell align="right" sx={cellSx}>
+                  {formatAmount(row.change)}
+                </TableCell>
+                <TableCell align="right" sx={cellSx}>
                   {row.balance && formatAmount(row.balance)}
-                </TableCell>{' '}
-                <TableCell align="right" sx={{ paddingLeft: 0 }}>
-                  <ExternalLink
+                </TableCell>
+                <TableCell
+                  align="right"
+                  sx={{ paddingLeft: 0, paddingRight: { xs: 2, md: 3 } }}
+                >
+                  <Link
                     href={txLink(getChain(config, row.chainId), row.txHash)}
-                  />
+                    target="_blank"
+                    rel="noopener noreferrer nofollow"
+                  >
+                    <FaArrowUpRightRegular sx={{ fontSize: 12 }} />
+                  </Link>
                 </TableCell>
               </TableRow>
             );
           })}
         </TableBody>
       </Table>
-    </TableContainer>
+      <TablePagination table={table} />
+    </Stack>
   );
 }
