+++ libs/oeth/history/src/queries.graphql
@@ -33,7 +33,12 @@ query HistoryTransaction($address: String!, $filters: [HistoryType!]) {
 }
 
 query Transfers ($tokens: [String!], $account: String!) {
-  erc20Transfers(where: {OR: {address_in: $tokens, from_containsInsensitive: $account}, address_in: $tokens, to_containsInsensitive: $account}) {
+  erc20Transfers(
+    orderBy: timestamp_DESC
+    offset: 0
+    limit: 2000
+    where: {OR: {address_in: $tokens, from_containsInsensitive: $account}, address_in: $tokens, to_containsInsensitive: $account}
+  ) {
     id
     chainId
     txHash
